Change order of evaluation in translation of local. Refs #15.
The generatio of C code for locals evaluates subexpressions before
generating the declaration of the variables defined by local. This
leads to declarations being reversed, since the declarations needed by
subexpressions are listed (written to a state) before the declaration
for the enclosing, local definition. In the presence of nested locals,
this leads to broken C code.

This commit changes the code generation so that the effects of the
translation of the subexpression are done after the variable is
defined.