Made tests deal with empty string in CSV output. Refs #30.
The the CSV driver output has an emptry string at the very end, just like all
the other lines have. The interpreter CSV output does not (a lack of a
newline?). By removing the last element of the driver output, this is fixed for
now. In the future, the interpreter should output more straightforward CSV, and
then we can revert this quick fix.